Fix deadlock issue in SatelliteController
The nested locks of mSatelliteEnabledRequestLock and
mIsSatelliteEnabledLock are causing a dead lock between main thread and
SatelliteController thread.
To avoid this, it shall not use the nested locks because it is easy to
cause a dead lock issue. With this change, move out of the lock scopes each other.

Change-Id: If6f714b5f1890e0713f22132026c8e26bdacc726
CRs-Fixed: 3966756